'ImageFinderClass': Constructor: added the option to use 'find_extension_files_subdir' instead (new function that also searches through subdirectories in the specified folder) Function find_extension_files_subdir: added. searches through subdirectories in the specified folder
added ImagePreprocessingClass: Intent is to resize and rename all (currently available) original bee images and store them also on the D drive (in the project folder). About 14000 images take less than 3GB of storage.

added testcode at the end (main fucntion): to test and execute the renaming and resizing of the original files