BackgroundImageClass: Fixes of tqdm usages. and an error with index (wrongly self.index).
ParentImageClass:
Expanded the constructor for arguments of the child classes (BeeFocusImage)
Added a DEBUG option for the constructor. (and processing functions: p00, p10,...)
Changed the generation of the open_close_kernel: now with fixed kernel and variable iterations (based on desired kernel size)
p10_diff_from_bg: The cast to uint8 is now done correctly inside this function.
p25_openclose_threshold: New function. We forgot to reduce the threshold image, before detecting contours. this is now fixed. ALSO, this function has an additional EROSION before, and DILATION after performing the CLOSING(2) and OPENING(2). This seems to result in better seperation of bee blobs.
p30_contours_extract_check: New function. Extracts contours from reduced image. Then the area of all contours is checked for min/max size. all fitting contours are added to the contour list.
process_1: New function. Executes the processes added so far.
In the constructor of ParentImageClass, this process_1 is executed and the list of childen (bee focus image) is filled from all valid contours.

BeeFocusImage:
Added calculation and getter function for the minAreaRect of the contour.
Added missing offset (of the focus image anchor) in calculating the outer contour (dilated mask).

Testfunctions for the new code.